backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
SQLAlchemy
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.2
pydantic-settings==2.2.1
python-dotenv==1.0.1
python-jose==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
requests==2.31.0
tldextract==5.1.1
httpx==0.26.0
python-multipart
celery
redis==5.0.1
sqlparse==0.5.2
mcp==0.9.1
openai==1.50.0